arch: Make up_systemreset() an arch global function
Previously, up_systemreset() was internal to the arm architecture.
In order to make it accessible to drivers, move the declaration
to the arch global header file, adding a new conditional so it is
only defined if the microprocessor supports it.

Change-Id: I5edb00f58592e7a54b416f90e252fb5a1d2d54e1
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/787717
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>