Ara: SVC: Implement hotplug state machine at boot time
At boot time store the hotplug states of the interfaces until the AP is ready.
When the AP is ready, consume the hotplug state and send hotplug
events to the AP.
Note: On interface hotplug, do not send the hotplug event to the AP
right away, wait for the handshake mechanism to succeed before doing so.

Testing Done: Check the hotplug state at boot and when the AP is ready
Bug Number: SW-82
Review URL: https://reviews.leaflabs.com/r/1710/
Reviewed by: fparent, mbolivar

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>