greybus: mods: datalink-spi: Improvement to WAKE line handling
The current version of the Silego part (which controls the WAKE, RFR,
and interrupt lines) has a limitation where if the RFR line is asserted,
the WAKE line is invalid (and therefore de-asserted). A future version
removes this limitation. The problem is the current code depended on
the WAKE line toggling for each packet. With this patch, if the base
sends multiple packets with the WAKE line asserted, it will be handled
correctly.

Change-Id: I8fc0ab4aa64c4abf45ee955c4af7023ccf0a450f
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/788096
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>