greybus/timestamps: add interface to capture tx/rx turnaround times
This patch adds an API to capture the difference between a TX and an RX
timestamp. And a mechanism to insert the graphed latency into reserved
fields in the greybus/loopback header. This will only be done for cport
indices with tag == true. Subsequent patches rely on this API so applying
this change first.

Testing Done: Tested with in-tandem with kernel changes
Bug Number: SW-1099
Review URL: https://reviews.leaflabs.com/r/1768
Reviewed by: Eli Sennesh, Fabien Parent, Bartosz Golaszewski
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>