switch: HACK: quick fix for power mode change race
The attribute TSB_DME_POWERMODEIND should be read to learn the results
of a UniPro power mode change on a given port. It normally reads zero,
but latches a sequence of different values as the power mode change
progresses. Software can read those values out, at which point the
attribute reverts to zero until the next change. Either polling or a
switch interrupt can be used to determine when the attribute has latched
a new value.

This means we have a race, as the power mode change code in
switch_apply_power_mode() polls the attribute after starting the power
mode change, but the IRQ worker for the ES2 switch also tries to read it
when it receives the interrupt.

Paper over that for now by disabling the read in the IRQ handler.

It's not clear how often we'd run into this race, but let's avoid it
anyway.

Testing Done: verify that power mode changes work on SDB
Bug Number: SW-1239
Review URL: https://reviews.leaflabs.com/r/1741/
Reviewed by: Mark Greer, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>