usb_test: fix and add error reporting
Currently, we are counting the number of transfer that failed but
we never report it to user.
Report the number of transfer that failed to user in case of error.
In addition, fix the data checks that was always detecting errors.

Testing Done: usb_test report errors if a least one transfer failed
Bug Number: SW-1007
Review URL: https://reviews.leaflabs.com/r/1753/
Reviewed by: fparent

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>