arm: stm32: Allow serial to use different clock source
By default, the PCLK is used as the U(S)ART source. With this patch,
boards can specify the clock frequency for each U(S)ART. If the
HSI clock is enabled and the specified frequency matches the HSI
frequency, the HSI clock source is automatically selected for the
board.

This will be useful for boards that need to use the HSI clock for
their U(S)ART. On the STM32L4x6, the U(S)ART can be a wakeup source
if the source clock is HSI, but not if using the PCLK.

Change-Id: Ib6d9427e1d4d9e32f4707222e8b817ec359093f3
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/789721
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: James Wylder <jwylder@motorola.com>