greybus/timestamps: capture firmware processing times
This patch provides a timestamp showing the amount of time it takes for the
GPBridge firmware to capture the UniPro RX interrupt, schedule the loopback
thread and go on to submit the corresponding response to the GPBridge
UniPro layer. The amount of time it takes to send that response over UniPro
is not included in the payload - since it cannot be. This patch provides
a view of the 'software overhead' involved only.

Testing Done: Tested with in-tandem with kernel changes
Bug Number: SW-1099
Review URL: https://reviews.leaflabs.com/r/1771
Reviewed by: Eli Sennesh, Bartosz Golaszewski, Mitch Tasman
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>