configs: hdk: muc: Set PCLK1 to HCLK frequency
PCLK1 is the PCLK for SPI2, which is used to communicate with the
base. The STM32L4x6 reference manual states that the SPI slave
frequency is up to fPCLK/2. The base is currently trying to communicate
at 4.8 MHz, which means the SPI block needs a PCLK of 9.6 MHz to
operate. The board's HCLK is currently 16 MHz, which means that
before this change, the PCLK was only 8 MHz. Bumping the PCLK to 16 MHz
guarantees SPI operation.

Change-Id: Ife30de116f794ae90b1f317ffb211f3a2d34889f
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/790573
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>