apps: slice: Enable PM and IDLE_CUSTOM / Disable IWDG
When the slice is not attached to a base, allow the STM32 to
enter Stop Mode to save power. The BPLUS interrupt will wake
the STM32 from Stop Mode when the slice is re-attached to a
base.

The common STM32 idle loop has bugs that prevent it from being
used when power management is enabled:
1) If a driver rejects the new power state, the power state is
still entered anyway.
2) The clocks aren't reset after exiting Stop Mode (however,
since the SPU is not currently using the PLL, this may be
optional).

The Internal Watchdog (IWDG) needs to be disabled, else it will
fire and reset the system when in Stop Mode. The watchdogs on
STM32 cannot be stopped once started.

Change-Id: Iae6914a4ebdedf3c9d5b6c54107ded5b87781719
Signed-off-by: Jim Wylder <jwylder@motorola.com>