DWC USB OTG: Change the buffer status before to complete transfer
Currently, if the status of a buffer is BS_DMA_DONE, then the
completion handler is called. The status should be updated to BS_HOST_BUSY.
It is required to avoid to call the completion handler more than once if
the completion handler take a while to transfer the data to UniPro (async).

Testing Done: The completion handler is only called once by request
Bug Number: SW-1006
Review URL: https://reviews.leaflabs.com/r/1856/
Reviewed by: mbolivar, fparent

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>