tsb: scm: add api to request/release pinshare access
In order to manage dynamically the pinsharing, we need to have each driver
request access to the PIN they need. This commit adds the function to request
and release a pinshare access.

The pinshare access settings are not enforced in this commit. It will be
enforce in a following commit.

Testing Done: gpio demo + tested all sequences of request/release to see if
              the api behave as expected.
Bug Number: SW-1223
Review URL: https://reviews.leaflabs.com/r/1819/
Reviewed by: mgreer, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>