tsb: scm: enforce pinsharing ownership
Until the port to the dynamic pinsharing code was complete the pinsharing
ownership was not enforced in order to have each seperate commit to still
work as expected. Now that the port is complete we can start to enforce the
pinsharing ownership and prevent drivers to set/clear pinshare of pin that
they don't own.

Testing Done: checked that trying to set the value of a pinshare of some pins
              fails if we don't own all the pins.
Bug Number: SW-1223
Review URL: https://reviews.leaflabs.com/r/1833/
Reviewed by: mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>