tsb: unipro: execute mailbox code only on mailbox interrupt
There is mailbox interrupt that can be generated that are using the unipro_irq
interrupt handler. This will soon not be the case anymore, so only execute
the mailbox code on mailbox interrupt and add an helper function to handle
the different types of events.

Testing Done: gpio demo
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2017/
Reviewed by: eligottlieb, gregkh, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>