tsb: i2s: Make AUDIOSET EDGE bit opposite for SO vs. SC and SI blocks
The Toshiba documentation is not at all clear about what the EDGE bit
in the AUDIOSET I2S register really does.  After examining examining
bad waveforms and some test code from another hardware company, it
appears that the bit must as follows:

a) If WCLK is to change on the rising edge of BCLK, then the EDGE bit in
   the SC and SI blocks should be set and the EDGE bit in the SO block
   should be cleared.

b) If WCLK is to change on the falling edge of BCLK, then the EDGE bit in
   the SC and SI blocks should be cleared and the EDGE bit in the SO block
   should be set.

It still isn't clear why this is the case from the documentation but this
"algorigthm" produces correct waveforms.

Testing Done: Ran i2s_test and examined I2S output with an analyzer
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/1883/
Reviewed by: fparent, mbolivar

Signed-off-by: Mark Greer <mgreer@animalcreek.com>