greybus: mods: attach: Switch to new dead base detection algorithm
Instead of relying on the Silego IC to detect the dead base, the
MuC must now detect a dead base on its own. A dead base will not
enable BPLUS to the mod, but will cause the SPI CS line to assert.
If the CS line is asserted for over 1 second, then the MuC knows
a dead base is attached.

In order to make this change possible, the driver was reworked to
allow other drivers to be able to register before the attach driver
is initialized. Attach initialization now consists of just interrupt
registration.

Change-Id: If8c9a05bc21546326d48d1aa80ba7bf57403b87a
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/796262
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Dmitriy Gruzman <dgruzman@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>