greybus: mods: datalink-spi: Clear RFR/INT if left active
Check if RFR is still set when we reach the txn_finished_cb before
txn_start_cb has a chance to clear it. If it is set, we'll explicitly
call back into the txn_start_cb to clear the RFR and INT lines. If this
is not cleared we'll get in a state where we'll no longer be able to
receive packets from the AP.

Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>

Change-Id: Idd83acd7493fabada00435fe6074c4a5e42638d9
Reviewed-on: http://gerrit.mot.com/797106
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>