svc: switch: es2: return value of requested switch register rather than 0
Fix es2_sys_ctrl_get() so that it writes the value of the requested switch
register into the result parameter, rather than 0.  Upon a successful get,
we had been writing the result code for the get operation rather than the
requested value.

A sequella of this bug is that we had not been executing the required M-PHY
fixup sequence for the ES2 switch correctly, since we read the M-PHY Trim
registers as 0, as opposed to the values programmed in eFuse.  As a result,
(though we do not understand the precise mechanism), the bit rate on the
M-PHY Lanes from Switch to Bridge in PWM Gear 1 was around 5.4 mbps rather
than the anticipated 8.5 mbps.

Testing Done: View content of the mphy_trim array in es2_fixup_mphy() before
    and after applying this fix, and verify that post-fix the values are
    non-zero.  Also, examine the Switch to APB2 link at PWM Gear 1 using a
    UniPro analyzer before and after applying this fix, and verify that
    after-fix link speed is as anticipated.
Bug Number: SW-1575
Review URL: https://reviews.leaflabs.com/r/1927/
Reviewed by: mgreer, fparent

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>