arm: stm32f446: don't register invalid CCM ram bank
The STM32F446 does not include the CCM RAM bank, but this bank is always
registered as heap even when the chip does not have it.
Check that the chip has the CCM ram bank before registering it.

This issue caused crashed in the SVC and erroneous -ENOMEM errors.

Testing Done: svc not crashing anymore.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1935/
Reviewed by: mbolivar, mgreer, jpihet

Signed-off-by: Fabien Parent <fparent@baylibre.com>