greybus: Add headroom support
The registered gb_transport_backend can now request headroom on the
greybus operation buffers. This can be used to prevent the reallocation
of a buffer and memcpy of the data just to add a header. The one caveat
is the headroom must be a multiple of four so the greybus buffer stays
aligned.

Change-Id: If40e7ec562763ea6a7e28f785eb2730521273d9f
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/798184
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>