apps: ara: arapm: improve readability of continuous mode
Clearing the screen in continuous mode make it hard to read the measurement
because of the blinking that happens. Instead of clearing the screen, rewind
the cursor and update what was previously printed on the screen.

Testing Done: tested on DB3
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1955/
Reviewed by: jpihet, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>