svc: fix improper mutex usage
Some mutexes are used in the following ways:
    1. pthread_mutex_lock(mutex_a);
    2.     pthread_mutex_lock(mutex_a);
    3.     some code
    4.     pthread_mutex_unlock(mutex_a);
    5. some code
    6. pthread_mutex_unlock(mutex_a);

Because the mutexes are not configured as recursive the second call to
pthread_mutex_lock (2) will fail because a deadlock will be detected. The
unlock line 4 is still called and thus the lock will be release on line 4
instead of line 6. The code on line 5 that shall be protected will actually
be called in an non-atomic manner.

This also break the usage of pthread_wait_cond() because this function will
be called when the mutex has been unlock when it shouldn't have been. This
will make pthread_wait_cond() fails all the time and it is breaking greybus.

Recursive mutexes cannot be used here because the mutex is also used with
a condition variable, so the fix is to check whether we were already holding
the mutex to know whether we should call pthread_mutex_unlock() or not.

This commit also fixes some missing/extra whitespaces.

Testing Done: gpio demo on db3
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1954/
Reviewed by: Mark Greer

From: Fabien Parent <fparent@baylibre.com>
Signed-off-by: Fabien Parent <fparent@baylibre.com>