apbridgea: fix possible use-after-free in cport reset
The code is assuming that everything will go well in CPort reset. If the
execution does not go as expected, we will encounter a use-after-free. Also
the code tries to send the response twice to the AP, this could also generate
some unexpected behaviour from the USB stack.

Testing Done: checked that the sequence for cport is reset is correct
Bug Number: SW-1459
Review URL: https://reviews.leaflabs.com/r/2012/
Reviewed by: eligottlieb, ptitiano

Signed-off-by: Fabien Parent <fparent@baylibre.com>