bridge: apbridgea: gadget: optimize memory usage
Use the bufram page allocation in order to reduce condiderably the memory
overhead of the buffer allocations in the usb gadget. With this change the
overhead is now 0% instead of being 100% as it was done before.

Testing Done: gpio demo
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2010/
Reviewed by: mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>