Ara: SVC: Switch: automatically re-enable ports interrupts when disabled
As suggested by Marti Bolivar.

The port interrupt can be disabled by the switch and/or the bridges.
In that case automatically re-enable the port interrupt. Since the interrupt
is latched in the switch it will fire again after enablement.

This fixes the multiple bridges hotplug, where the mailbox interrupt
was not seen by the SVC.

Testing Done: Multiple hotplug of the bridges on SDB
Bug Number: SW-1527
Review URL: https://reviews.leaflabs.com/r/1987/
Reviewed by: Fabien Parent, Eli Sennesh

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>