ara/bridge: for OOT module support, include C files instead of object files
This patch includes C files in the Makefile for OOT board support instead of
receiving already compiled object files from the FDK. It thus takes advantages
of the existing rules for compiling these C files into objects. We are also able
to inherit the same compilation options as the other object files.

This patch will also simplify the Makefile in the FDK since we will be able to
remove the rules that compile the C files related to OOT board support in
objects (and which are merely a copy of the rules already included in this
Makefile).

Testing Done: compiled and tested with gpio demo
Bug Number: SW-1708
Review URL: https://reviews.leaflabs.com/r/1991/
Reviewed by: fparent, mbolivar

Signed-off-by: Joel Porquet <joel@porquet.org>