ara: add post_ara_patches.py
This is a script that makes it much more convenient to post ReviewBoard
patch series.

Call it with a revision range:

  post_ara_patches.py rev1..rev2

To post the patches in the range, in order, with the proper "depends on"
fields filled out in ReviewBoard so people can read the patch series
more easily.

Testing Done: posted a multipatch series using the script
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2055/
Reviewed by: Fabien Parent, Mark Greer

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>