DWC USB OTG: Only use the queue ring for bulk out endpoints
Due to the complexity of USB driver, the endpoint queue ring
may be used by bulk in endpoint in wrong way.
Following patches are going to use the queue ring to store
the pending requests that will be used to program DMA.
Ensure we are only using a queue ring for bulk out endpoints,
to have a clean base for the following patches.

Testing Done: Tested with gb_loopback
Bug Number: SW-1704 && SW-1006
Review URL: https://reviews.leaflabs.com/r/2148/
Reviewed by: mbolivar

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>