DWC USB OTG: Update dwc_otg_pcd_queue_req() to return an error if it fails
dwc_otg_pcd_queue_req() is to queue a request to sg_dma_queue
and to reprogram the DMA with this request if the request is new.
But the function may fails to program the DMA if the endpoint is
enabled.

Testing Done: Tested with gb loopback and more than bulk out request
Bug Number: SW-1006
Review URL: https://reviews.leaflabs.com/r/2158/
Reviewed by: mbolivar, gregkh

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>