version: add binary to all builds
We lost the version binary in the build as it was not hooked up in
apps/examples/Kconfig so the next time the defconfigs were updated, the
changes were removed.

So hook up version to the correct Kconfig chain, make it 'default y' so
that all binaries get it (hint, we always want it there), and add it to
all of the defconfigs as well.

Testing Done: all configs built with it enabled
Bug Number: N/A
Review URL: https://reviews.leaflabs.com/r/2175/
Reviewed by: Marti Bolivar, Fabien Parent

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>