i2c: move i2c greybus defines to i2c-gb.h
The I2C Greybus defines should all go in the i2c-gb.h, not some in the
.c file, and others in the .h file.  Also flush out the defines to
include all of them as specified in the greybus spec.  We will need
these in a later patch, and add a "GB_" prefix to them so they don't
conflict with Nuttx defines that might have the same name, but a
different value.

Testing Done: all configs built
Review URL: https://reviews.leaflabs.com/r/2177/
Reviewed by: Fabien Parent

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>