DWC USB OTG: Fix the descriptor counter
To differentiate the use ring DMA descriptors chain or the regular DMA descriptor chain,
we are using dwc_ep.desc_cnt. For this variable, 0 mean regular DMA descriptor chain.
We are always reseting this variable to zero and if we want to use the ring DMA descriptor chain,
then we have to set it.
But this variable is also used to count the number of descriptors. It's responsability of
ring DMA methods to correctly set it but sometime, the we can't recover this value.
Add second  variable to store the number of descriptors and restore dwc_ep.desc_cnt.

Testing Done: Tested with gb loopback and more than bulk out request
Bug Number: SW-1006
Review URL: https://reviews.leaflabs.com/r/2184/
Reviewed by: gregkh

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>