tsb: scm: fix clock definitions
In ES2 the unipro_ref clock disappeared, and with it disapearing the clock bit
mapping in the clock gating register 1 changed. The definitions were not changed
when porting NuttX to ES2 and this did not create any issue since on ES2 all
the clocks are enabled after boot.

On ES3, HSIC clocks are disabled after boot which means we were not setting the
correct value when clocking USB.

This issue was discovered by Mitch Tasman.

Testing Done: gpio demo on ES2 and ES3.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2218/
Reviewed by: mtasman, morgang

Signed-off-by: Fabien Parent <fparent@baylibre.com>