greybus: mods: datalink-spi: Reduce length of log prints
Function names are automatically added to debug prints. By
reducing the length of function names, the length of the debug
prints in that function are reduced.

The goal of this change is to reduce the extraneous logging so
more prints can fit in the last dmesg buffer and aid in debugging
issues.

Change-Id: I24f06ce390992b6c2d5bcecf90d702f81ea7d51c
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/812596
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: James Wylder <jwylder@motorola.com>