greybus: mods: datalink-spi: Fix attach_cb & wake_isr race condition
On mod removals, a race condition is seen between attach_cb and
the wake_isr. The attach_cb would be called first, and be in the
middle of cleaning up when the wake_isr would fire due to a false
wake signal, setting up for a transfer. This results in an unknown
and invalid state, and communication with the base is no longer
possible on the subsequent attaches.

Now, attach_cb disables IRQs during cleanup, so the wake_isr fires
after the cleanup and reads the correct state of detached.

Change-Id: I29e4df3ba02fc6c3ffb5d7577c46748b11aa07a8
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/813676
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>