arm: stm32: rtc: Adjust RTC calibration when setting time
Add a simple algorithm for fine tuning the RTC clock speed when
the RTC time is set. Basically, if the new RTC time is ahead of
the current time, the clock speed is increased slightly and vice
versa.

However, if the delta is too large, it is assumed that the time
has actually changed (as opposed to a clock sync) and the clock
rate is not adjusted. Likewise, if the time delta is very small,
then consider the clock rate "good enough" and do not adjust it.

Change-Id: Iebb32bfa4c1a23a7d742540cb7b21168910c79b8
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: https://gerrit.mot.com/896076
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>
Reviewed-by: Makarand Karvekar <mkarvekar@motorola.com>