greybus: manifests: hdk: Split into multiple bundles
Recently, there was a change to the greybus code on the base side
to tear down the entire bundle if one of the cports in that bundle
fails. Because of this, split bundle 1 into multiple bundles.
This way, if there is a failure on one of the cports, only that
specific bundle is affected.

The HID cport was removed because the new HDK no longer has HID
button support.

Change-Id: Ib7eedaf12417ffd03b22d69f6c693f4e3c865f64
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/818817
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>