greybus: mods: datalink-spi: Add support for first packet hdr bit
The SPI protocol has been updated to add a bit to the bitmask in
the header to indicate whether the packet is the first packet.
This allows for any packet loss to be detected. If the first
packet is lost of a multi-packet message, the next packet will
not have the first packet bit set, so the driver knows to drop
the remaining packets.

Change-Id: I0a599e524002424cf788f4b8c4a8bc59916b26ff
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/826020
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>