syslog: Move current ramlog to ramlog section of memory
In addition of the last ramlog, the current ramlog also needs to
be in the ramlog section of memory. This was needed because on
boards with a bootloader, the bootloader was clobbering the ramlog
memory, making it unavailable on the next Nuttx boot to be saved
as the last ramlog. By moving the current ramlog into the same
section as the last ramlog, the bootloader can be setup to avoid
this known section of memory, thereby preserving the ramlog.

Because this is just moving RAM, there is no additional RAM usage.

Change-Id: I1f4dff8d4b5989484acc2e87f9ebf42817172ed2
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/835606
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>