tsb: Add free running timer driver code for the TSB
A new free running timer driver allows different low level drivers
to access a free running version of the 48MHz timer.  This allows
for unified power management allowing the clock to be shutdown when
no free running timer is in use.  Please be aware that at this time
this is not integrated with the free running timer used for the
tickless version of the kernel.

Change-Id: I90e423af75fab4b0029ed221711e87ede0341aad
Signed-off-by: David Peyer <David.Peyer@motorola.com>
Reviewed-on: http://gerrit.mot.com/835532
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: David Peyer <wldp07@motorola.com>
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: Kazuhiro Ondo <kazuhiro.ondo@motorola.com>