greybus: mods: datalink-spi: Do not ACK dummy packets
To minimize the chance that a valid packet could be sent and
falsely ACK'd, do not ACK when a dummy packet is received. The
core does not look for an ACK when it sends a dummy packet
anyway, so ACK'ing it only increases the odds that a false
ACK could be seen in a perfect storm scenario.

Change-Id: I4d428bcfa18c192c56bcf282a20464db474cebad
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/840842
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>