arm: stm32: spi: Throttle error prints
When SPI errors occur, the first errors are important for analyzing
how we got into the bad state. If many errors happen in a row,
the log is flooded and the beginning of the log buffer is lost.

Because repeating errors tend to be the same, throttle the error
prints to only print the first five errors. A successful transfer
resets the counter.

Change-Id: Id4bc82665f93e057218e48430fb654f482619fd7
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/843271
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>