configs: hdk: Discontinue use of Low-Power Run
It turns out Low-Power Run mode isn't very useful for us. We could
not communicate via SPI at the slower clock speed, so when the core
wanted to send us a message when the MuC was in LPR, it had to wait
the lengthy delay to get the clock cranked back up. The greatly
increased the latency for this use case. In addition, periodically
watchdog resets were seen because the watchdog pets came too slowly
when in LPR.

Because of this, there is no real reason to go to LPR. Change the
IDLE mode to be the same as NORMAL mode, and get to STOP mode faster.

In short, the goal now is to wake up, do what is required, and get
back to STOP quickly.

Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/785318
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>
Submit-Approved: James Wylder <jwylder@motorola.com>

Change-Id: I93d84759d88927172edca2ae4027deb30cf4beb4