ramlog: Use new .ramlog_dev section for ramlog_dev_s
With multiple variables in the .ramlog section, the compiler is
not guaranteed to place them in a particular order. In order to
support an external app (most likely a bootloader) to access the
ramlog, the ramlog_dev_s must always be in a known location. To
accomplish this, the ramlog_dev_s is placed in a new .ramlog_dev
section so it can be accurately placed in the ld.script.

Change-Id: Ia033699b3e429f3944b04120d70ed2e335d835ca
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/845925
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: James Wylder <jwylder@motorola.com>