nuttx: drivers: mhb: dsi_display: Read panel info
Read the panel info and save it in the display structure.  The
panel info is not currently used to select the config structure.

Change the states to an enum to reduce the risk of mistakes.

Change-Id: Ie982f3b4161bfb05c7ee625292d8e384408b4ff3
Signed-off-by: Mike Corrigan <corrigan@motorola.com>
Reviewed-on: http://gerrit.mot.com/845634
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Carl Grube <cgrube@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Kazuhiro Ondo <kazuhiro.ondo@motorola.com>