apps: mods: diet: Move attach init after cport registrations
In addition to needing to be initialized after the network, the
attach driver initialization should also be after the cport
registrations. This solves the use case where the user attaches
a dead mod or mod with no battery to a base. The base will enable
BPLUS, which causes the mod to power up and boot Nuttx. The base
will then attempt to communicate with the mod. Previously the mod
would start to communicate before all the cports were registered,
resulting in false errors being returned to the base.

By moving the attach after the cport registrations, the mod will
not communicate until it is ready.

Change-Id: I1da20b59738a2bf012ec1ca27fcf9e262d1080e5
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/847231
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>