manifests: hdk: Simplify manifests
Previously the manifests had many cports for many features to
show what can be done with mods. To aid with the learning curve,
redo the manifests to have a minimal base and have specific
examples build from there.

hdk.mnfs: The base HDK board with no personality card. The bare
minimum to get up and running. Mods can be made on top of this.

hdk-display: Takes the base hdk.mnfs and adds support for
DisplayPort.

hdk-mhb: Takes the base hdk.mnfs and adds MHB support. (Not
included in this patch as it was created with the minimal cports
required to demonstrate this feature.)

Change-Id: If542419538c828357c2c0218345703c71dd8e5ab
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/847746
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Michael Corrigan <michael.corrigan@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>