configs: hdk: muc: Add blinky LED raw driver
The blinky LED raw driver is a (hopefully) simple example to
demonstrate how the raw protocol can be used to run a task.

When a non-zero message is received over the raw protocol, an
LED is blinked. When zero is received, then stop the blinking.
A hardware timer is used to control the blink.

Change-Id: I515d8416e8230a2d571b12f616427616289a17d0
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/849168
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>