device_pwr_limit: Add Current Reservation Message
Add message to allow the mod to notify the base that
they are going to use the full power (above 1A). This
should only be sent when the pwr_limit is set above
the 1A limit.  If the mod is using the higher limit
it should wait until the current_rsv_ack message
is received before pulling the full current.  If
release the current reservation, it is expected the
mod is already lowered it's usage.

The core should not reject this reservation, it has
already granted the power limit though other means.
This is only a notification.

Change-Id: I004d8105ded513c66ead14bd7d1fcfd6bc342a70
Signed-off-by: Jim Wylder <jwylder@motorola.com>
Reviewed-on: http://gerrit.mot.com/850649
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Gregory Meiste <w30289@motorola.com>