arch: arm: tsb: cdsi: Calculate PLL if not specified
If the PLL settings are not specified in the CDSI config
structure, then calculate the best PLL settings (i.e.
the valid settings with the least difference between
the target and actual rate).

It may be that these settings are not 'good enough' in
practice, but it is the best we can do.

Refactor the PLL validation and hsck calculation into
it's own function to support the PLL calculation logic.

Change-Id: I2c774b76f1790681b2083ce4d702dbd3c81b07af
Signed-off-by: Mike Corrigan <corrigan@motorola.com>
Reviewed-on: http://gerrit.mot.com/854096
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Kazuhiro Ondo <kazuhiro.ondo@motorola.com>